+++ sampledb/api/server/files.py
@@ -4,16 +4,18 @@ RESTful API for SampleDB
 """
 
 import base64
-import flask
 
+import flask
 from flask_restful import Resource
 
 from .authentication import object_permissions_required, Permissions
-from ...logic.actions import get_action
-from ...logic.objects import get_object
-from ...logic.files import File, get_file_for_object, get_files_for_object, create_local_file, create_url_file, create_database_file
-from ...logic.utils import parse_url
 from ...logic import errors
+from ...logic.actions import get_action
+from ...logic.files import create_database_file, create_local_file, \
+    create_local_file_reference, create_url_file, File, get_file_for_object, \
+    get_files_for_object
+from ...logic.objects import get_object
+from ...logic.utils import parse_url
 
 __author__ = 'Florian Rhiem <f.rhiem@fz-juelich.de>'
 
@@ -83,9 +85,9 @@ class ObjectFiles(Resource):
                 "message": "storage must be set"
             }, 400
         storage = request_json['storage']
-        if storage not in ('local', 'url', 'database'):
+        if storage not in ('local', 'local_reference', 'url', 'database'):
             return {
-                "message": "storage must be 'local', 'database' or 'url'"
+                "message": "storage must be 'local', 'local_reference', 'database' or 'url'"
             }, 400
         if storage in {'local', 'database'}:
             for key in request_json:
@@ -123,6 +125,26 @@ class ObjectFiles(Resource):
                     file_name=original_file_name,
                     save_content=lambda stream: stream.write(content)
                 )
+        if storage == 'local_reference':
+            for key in request_json:
+                if key not in {'object_id', 'storage', 'filepath'}:
+                    return {
+                        "message": "invalid key '{}'".format(key)
+                    }, 400
+            if 'filepath' not in request_json or not request_json['filepath']:
+                return {
+                    "message": "filepath must be set for files with local_reference storage"
+                }, 400
+            try:
+                file = create_local_file_reference(
+                    object_id=object_id,
+                    user_id=flask.g.user.id,
+                    filepath=request_json['filepath']
+                )
+            except errors.UnauthorizedRequestError:
+                return {
+                    "message": "user not authorized to add this path"
+                }, 403
         if storage == 'url':
             for key in request_json:
                 if key not in {'object_id', 'storage', 'url'}: